% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cattle.R
\name{add_cattle}
\alias{add_cattle}
\title{Add cattle data to the AnitaCattle collection in the PLMResearch database.}
\usage{
add_cattle(
  RFID,
  mtag,
  calvingdate,
  paddock,
  date,
  hour,
  status,
  username = username,
  password = password
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{mtag}{a list of cattle management tag number/s}

\item{calvingdate}{a list of cattle calving dates}

\item{paddock}{the paddock of the cow}

\item{date}{a list of dates}

\item{hour}{a list of cattle timestamps corresponding to the dates}

\item{status}{the calving status of the cow}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds cattle data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
