% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_status.R
\name{add_status}
\alias{add_status}
\title{Add status data to the AnitaStatus collection in the PLMResearch database.}
\usage{
add_status(
  RFID,
  mtag,
  calvingdate,
  timestamp,
  roundedtime,
  status,
  username = user,
  password = pass
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{mtag}{a list of cattle management tag number/s}

\item{calvingdate}{a list of cattle calving dates}

\item{timestamp}{a list of cattle timestamps of a coordinate point}

\item{roundedtime}{a list of cattle timestamps, rounded to the closest 5 minutes}

\item{status}{the calving status of the cow}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds status data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
