% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addaccelerometer.R
\name{addaccelerometer}
\alias{addaccelerometer}
\title{Add accelerometer data to the AnitaAccelerometer collection in the PLMResearch database.}
\usage{
addaccelerometer(
  RFID,
  mtag,
  date,
  hour,
  roundedtime,
  timestamp,
  X,
  Y,
  Z,
  status,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{mtag}{a list of cattle management tag number/s}

\item{date}{a list of cattle dates}

\item{hour}{a list of cattle hours}

\item{roundedtime}{a list of cattle rounded times}

\item{timestamp}{a list of cattle timestamps}

\item{X}{the x axes}

\item{Y}{the Y axes}

\item{Z}{the Z axes}

\item{status}{the status}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds accelerometer data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
