% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcalvingmodeldata.R
\name{addcalvingmodeldata}
\alias{addcalvingmodeldata}
\title{Add data that feeds into the calvingmodel function to the AnitaCalvingModel collection in the PLMResearch database.}
\usage{
addcalvingmodeldata(
  RFID,
  mtag,
  date,
  hour,
  hour.mean.MI,
  hour.mean.diff.Z,
  hour.max.diff.Z,
  hour.max.MV,
  self.hour.min.diff.all,
  self.hour.min.MV,
  self.hour.min.diff.entropy,
  self.hour.sd.diff.all,
  self.hour.range.X,
  self.hour.range.energy,
  self.hour.range.diff.X,
  self.hour.range.diff.Z,
  self.hour.range.diff.all,
  self.hour.range.MV,
  day.mean.X,
  day.max.diff.Y,
  day.max.diff.Z,
  day.max.MV,
  herd.day.mean.Z,
  herd.day.mean.all,
  herd.day.mean.MI,
  herd.day.mean.entropy,
  herd.day.min.diff.Y,
  herd.day.min.diff.Z,
  herd.day.max.diff.X,
  herd.day.sd.X,
  herd.day.sd.Y,
  herd.day.sd.Z,
  herd.day.sd.all,
  herd.day.sd.energy,
  herd.day.sd.diff.Y,
  herd.day.range.X,
  herd.day.range.diff.all,
  self.day.mean.X,
  self.day.mean.Y,
  self.day.mean.all,
  self.day.mean.MI,
  self.day.mean.diff.X,
  self.day.min.diff.Y,
  self.day.min.diff.Z,
  self.day.min.diff.all,
  self.day.range.diff.Y,
  self.day.range.diff.Z,
  MCP.herd.day.MCP,
  MCP.self.day.MCP,
  MMM.day.mean.dist,
  MMM.day.mean.diff.dist,
  MMM.self.day.mean.dist,
  MMM.self.day.sd.diff.speed,
  nearest.day.mean.diff,
  nearest.day.min.diff,
  nearest.day.max.diff,
  nearest.self.day.sd.diff,
  hour.rain,
  hour.humidity,
  hour.THI,
  day.minrain,
  day.rangerain,
  day.maxhumidity,
  day.meanhumidity,
  day.minTHI,
  day.meanTHI,
  day.sdTHI,
  username = user,
  password = pass
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{mtag}{a list of cattle management tag number/s}

\item{date}{a list of cattle dates}

\item{hour}{a list of cattle hours}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds data that feeds into the calvingmodel function from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
