% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcattle.R
\name{addcattle}
\alias{addcattle}
\title{Add cattle data to the AnitaCattle collection in the PLMResearch database.}
\usage{
addcattle(
  RFID,
  mtag,
  calvingdate,
  calving,
  paddock,
  date,
  hour,
  status,
  username = user,
  password = pass
)
}
\arguments{
\item{RFID}{a list of cattle RFIDs}

\item{mtag}{a list of cattle management tags}

\item{calvingdate}{a list of cattle calving dates}

\item{calving}{a list of calving styles}

\item{paddock}{a list of cattle paddocks}

\item{date}{a list of dates for the data}

\item{hour}{a list of hours for the data}

\item{status}{a list of cow statuses for the given timestamp}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds cattle data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
