% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addgps.R, R/addsvmgps.R, R/addsvmweather.R
\name{addgps}
\alias{addgps}
\title{Add GPS data to the AnitaGPS collection in the PLMResearch database.}
\usage{
addgps(
  RFID,
  mtag,
  date,
  hour,
  period,
  hour.rain,
  hour.humidity,
  hour.THI,
  day.minrain,
  day.rangerain,
  day.maxhumidity,
  day.meanhumidity,
  day.minTHI,
  day.meanTHI,
  day.sdTHI,
  status,
  username = user,
  password = pass
)

addgps(
  RFID,
  mtag,
  date,
  hour,
  period,
  hour.rain,
  hour.humidity,
  hour.THI,
  day.minrain,
  day.rangerain,
  day.maxhumidity,
  day.meanhumidity,
  day.minTHI,
  day.meanTHI,
  day.sdTHI,
  status,
  username = user,
  password = pass
)

addgps(
  RFID,
  mtag,
  date,
  hour,
  period,
  hour.rain,
  hour.humidity,
  hour.THI,
  day.minrain,
  day.rangerain,
  day.maxhumidity,
  day.meanhumidity,
  day.minTHI,
  day.meanTHI,
  day.sdTHI,
  status,
  username = user,
  password = pass
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{mtag}{a list of cattle management tag number/s}

\item{date}{the date of the coordinate point}

\item{hour}{the hour of the coordinate point}

\item{period}{the period}

\item{status}{the calving status of the cow}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}

\item{roundedtime}{the timestamp of the coordinate point, rounded to the closest 5 minutes}

\item{timestamp}{the timestamp of the coordinate point}

\item{latitude}{the latitude of the coordinate point}

\item{longitude}{the longitude of the coorindate point}
}
\value{
a message that indicates the data has been successfully added

a message that indicates the data has been successfully added

a message that indicates the data has been successfully added
}
\description{
This function adds GPS data from Anita's Belmont trial to the MongoDB database.

This function adds GPS data from Anita's Belmont trial to the MongoDB database.

This function adds GPS data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}

Anita Chang \email{a.chang@cqu.edu.au}

Anita Chang \email{a.chang@cqu.edu.au}
}
