% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnotes.R
\name{addnotes}
\alias{addnotes}
\title{Add cattle data to the AnitaCattle collection in the PLMResearch database.}
\usage{
addnotes(
  RFID,
  calfID = NULL,
  calfsex = NULL,
  calfvigour = NULL,
  calfwt = NULL,
  udder = NULL,
  front = NULL,
  back = NULL,
  cowbcs = NULL,
  notes = NULL,
  username = user,
  password = pass
)
}
\arguments{
\item{RFID}{a list of cattle RFIDs}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}

\item{mtag}{a list of cattle management tags}

\item{calvingdate}{a list of cattle calving dates}

\item{calving}{a list of calving styles}

\item{paddock}{a list of cattle paddocks}

\item{date}{a list of dates for the data}

\item{hour}{a list of hours for the data}

\item{status}{a list of cow statuses for the given timestamp}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds cattle data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
