% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpaddocks.R
\name{addpaddocks}
\alias{addpaddocks}
\title{Add paddock data to the AnitaPaddocks collection in the PLMResearch database.}
\usage{
addpaddocks(paddname, hectares, cattle, username = user, password = pass)
}
\arguments{
\item{paddname}{the name of the paddock}

\item{hectares}{the hectares of the paddock}

\item{cattle}{the number of cattle in the paddock}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds paddock data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
