% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addsvmaccelerometer.R
\name{addsvmaccelerometer}
\alias{addsvmaccelerometer}
\title{Add GPS data to the AnitaGPS collection in the PLMResearch database.}
\usage{
addsvmaccelerometer(
  RFID,
  mtag,
  date,
  hour,
  period,
  hour.mean.MI,
  hour.mean.diff.Z,
  hour.max.diff.Z,
  hour.max.MV,
  self.hour.min.diff.all,
  self.hour.min.MV,
  self.hour.min.diff.entropy,
  self.hour.sd.diff.all,
  self.hour.range.X,
  self.hour.range.energy,
  self.hour.range.diff.X,
  self.hour.range.diff.Z,
  self.hour.range.diff.all,
  self.hour.range.MV,
  day.mean.X,
  day.max.diff.Y,
  day.max.diff.Z,
  day.max.MV,
  herd.day.mean.Z,
  herd.day.mean.all,
  herd.day.mean.MI,
  herd.day.mean.entropy,
  herd.day.min.diff.Y,
  herd.day.min.diff.Z,
  herd.day.max.diff.X,
  herd.day.sd.X,
  herd.day.sd.Y,
  herd.day.sd.Z,
  herd.day.sd.all,
  herd.day.sd.energy,
  herd.day.sd.diff.Y,
  herd.day.range.X,
  herd.day.range.diff.all,
  self.day.mean.X,
  self.day.mean.Y,
  self.day.mean.all,
  self.day.mean.MI,
  self.day.mean.diff.X,
  self.day.min.diff.Y,
  self.day.min.diff.Z,
  self.day.min.diff.all,
  self.day.range.diff.Y,
  self.day.range.diff.Z,
  status,
  username = user,
  password = pass
)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{mtag}{a list of cattle management tag number/s}

\item{date}{the date of the coordinate point}

\item{hour}{the hour of the coordinate point}

\item{status}{the calving status of the cow}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds GPS data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
