% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addwow.R
\name{addwow}
\alias{addwow}
\title{Add walk over weigh data to the AnitaWoW collection in the PLMResearch database.}
\usage{
addwow(
  RFID,
  mtag,
  date,
  time,
  timestamp,
  weight,
  username = user,
  password = pass
)
}
\arguments{
\item{timestamp}{a list of timestamps for weather}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}

\item{rain}{the precipitation}

\item{temp}{the temperature (T)}

\item{humidity}{the humidity (H)}

\item{THI}{the temperature-humidity index. Calculated as: 0.8×T+H×(T-14.4)+46.4}

\item{condition}{the weather conditions}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds weather data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
