% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calvingmodel.R
\name{calvingmodel}
\alias{calvingmodel}
\title{Add accelerometer data to the AnitaAccelerometer collection in the PLMResearch database.}
\usage{
calvingmodel(RFID, mtag, username = user, password = pass)
}
\arguments{
\item{RFID}{a list of cattle RFID number/s}

\item{mtag}{a list of cattle management tag number/s}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds accelerometer data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
