% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cattle.R
\name{get_cattle}
\alias{get_cattle}
\title{Retrieves cattle data from the AnitaGPS collection in the PLMResearch database.}
\usage{
get_cattle(date, username = NULL, password = NULL)
}
\arguments{
\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}

\item{timestamp}{a list of timestamps}
}
\value{
returns the cattle spatial information
}
\description{
This function retrieves cattle data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
