% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gps.R
\name{get_gps}
\alias{get_gps}
\title{Retrieves GPS data from the AnitaGPS collection in the PLMResearch database.}
\usage{
get_gps(roundedtime, status, username, password)
}
\arguments{
\item{status}{the status of the cow}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}

\item{timestamp}{a list of timestamps}
}
\value{
returns the cattle spatial information
}
\description{
This function retrieves GPS data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
