% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_weather}
\alias{get_weather}
\title{Retrieves weather data from the AnitaWeather collection in the PLMResearch database.}
\usage{
get_weather(timestamp = NULL, username = NULL, password = NULL)
}
\arguments{
\item{timestamp}{a list of timestamps}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
returns the paddock spatial information
}
\description{
This function retrieves weather data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
