% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getaccelerometer.R
\name{getaccelerometer}
\alias{getaccelerometer}
\title{Retrieves cattle data from the AnitaCattle collection in the PLMResearch database.}
\usage{
getaccelerometer(time = NULL, username = user, password = pass)
}
\arguments{
\item{time}{a list of timestamps}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
returns the cattle spatial information
}
\description{
This function retrieves cattle data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
