% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmlresults.R
\name{getmlresults}
\alias{getmlresults}
\title{Add accelerometer data to the AnitaAccelerometer collection in the PLMResearch database.}
\usage{
getmlresults(hour, username = user, password = pass)
}
\arguments{
\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds accelerometer data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
