% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpaddocks.R
\name{getpaddocks}
\alias{getpaddocks}
\title{Retrieves paddock data from the AnitaPaddocks collection in the PLMResearch database.}
\usage{
getpaddocks(username = user, password = pass)
}
\arguments{
\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
returns the paddock spatial information
}
\description{
This function retrieves paddock data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
