% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweather.R
\name{getweather}
\alias{getweather}
\title{Retrieves weather data from the AnitaWeather collection in the PLMResearch database.}
\usage{
getweather(timestamp = NULL, username = user, password = pass)
}
\arguments{
\item{timestamp}{a list of timestamps}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
returns the paddock spatial information
}
\description{
This function retrieves weather data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
