% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwow.R
\name{getwow}
\alias{getwow}
\title{Retrieves weather data from the AnitaWeather collection in the PLMResearch database.}
\usage{
getwow(time = NULL, username = user, password = pass)
}
\arguments{
\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}

\item{timestamp}{a list of timestamps}
}
\value{
returns the paddock spatial information
}
\description{
This function retrieves weather data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
