% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{Retrieves weather data from the AnitaWeather collection in the PLMResearch database.}
\usage{
reset(username = user, password = pass)
}
\arguments{
\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}

\item{status}{the status}
}
\value{
returns the paddock spatial information
}
\description{
This function retrieves weather data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
