% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{training}
\alias{training}
\title{Add accelerometer data to the AnitaAccelerometer collection in the PLMResearch database.}
\usage{
training(username = NULL, password = NULL)
}
\arguments{
\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}

\item{RFID}{a list of cattle RFID number/s}

\item{mtag}{a list of cattle management tag number/s}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds accelerometer data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
