% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weather.R
\name{add_weather}
\alias{add_weather}
\title{Add weather data to the AnitaWeather collection in the PLMResearch database.}
\usage{
add_weather(
  timestamp,
  rain,
  temp,
  humidity,
  THI,
  condition,
  username,
  password
)
}
\arguments{
\item{timestamp}{a list of timestamps for weather}

\item{rain}{the precipitation}

\item{temp}{the temperature (T)}

\item{humidity}{the humidity (H)}

\item{THI}{the temperature-humidity index. Calculated as: 0.8×T+H×(T-14.4)+46.4}

\item{condition}{the weather conditions}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
a message that indicates the data has been successfully added
}
\description{
This function adds GPS data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
