% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gps.R
\name{get_gps}
\alias{get_gps}
\title{Retrieves GPS data from the AnitaGPS collection in the PLMResearch database.}
\usage{
get_gps(timestamp = NULL, status = NULL, username = NULL, password = NULL)
}
\arguments{
\item{timestamp}{a list of timestamps}

\item{status}{a list of cattle statuses to search for}

\item{username}{username for use with Anita's App}

\item{password}{password for use with Anita's App}
}
\value{
returns the cattle spatial information
}
\description{
This function retrieves GPS data from Anita's Belmont trial to the MongoDB database.
}
\author{
Anita Chang \email{a.chang@cqu.edu.au}
}
